// src/app/interfaces/branch.ts
export interface Branch {
    _id?: string;
    primaryPhoneNo?: string;
    secondaryPhoneNo?: string;
    addressLineOne?: string;
    addressLineTwo?: string;
    addressLines?: string[];
    email?: string;
    shopOpenTime?: string;
    notification?: string;
    type?: string;
    priority?: number;
  }
  